+++ lib/ep/person.rb
@@ -1,4 +1,6 @@
 # frozen_string_literal: true
+require_relative '../person_proxy_images'
+
 module EP
   class Person
     def initialize(person:, term:, mapit:, baseurl:)
@@ -12,17 +14,7 @@ module EP
     def_delegators :@person, :id, :name, :birth_date, :image, :phone,
                              :email, :twitter, :facebook, :memberships
 
-    def thumbnail_image_url
-      proxy_image_variant(:thumbnail)
-    end
-
-    def medium_image_url
-      proxy_image_variant(:medium)
-    end
-
-    def original_image_url
-      proxy_image_variant(:original)
-    end
+    include PersonProxyImages
 
     def twitter_display
       "@#{twitter}" if twitter
@@ -77,12 +69,6 @@ module EP
 
     attr_reader :person, :term, :mapit, :baseurl
 
-    ALLOWED_IMAGE_SIZES = {
-      thumbnail: '100x100',
-      medium: '250x250',
-      original: 'original'
-    }
-
     def ep_area
       current_memberships.first.area
     end
@@ -91,17 +77,9 @@ module EP
       term.legislature
     end
 
-    def proxy_image_variant(size)
-      return if image.nil?
-      raise_unless_image_size_available(size)
+    def proxy_image_base_url
       'https://theyworkforyou.github.io/shineyoureye-images' \
-      "/#{legislature.slug}/#{id}/#{ALLOWED_IMAGE_SIZES[size]}.jpeg"
-    end
-
-    def raise_unless_image_size_available(size)
-      unless ALLOWED_IMAGE_SIZES.has_key?(size)
-        raise "Size #{size} is not known to be available"
-      end
+      "/#{legislature.slug}/#{id}/"
     end
   end
 end
